requirements.txt
# Core Data Science Libraries
jupyter==1.0.0
jupyterlab==4.0.0
notebook==7.0.0

# Data Manipulation and Analysis
pandas==2.1.0
numpy==1.24.0

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.15.0

# Machine Learning
scikit-learn==1.3.0

# MCP Server Dependencies
arxiv==2.1.0
mcp==1.0.0
fastmcp==0.9.0

# Azure OpenAI Client Dependencies
openai==1.12.0
python-dotenv==1.0.0
nest-asyncio==1.5.8

# Additional Utilities
requests==2.31.0

# Development Tools
black==23.7.0
flake8==6.0.0
pytest==7.4.0
